CF-24 move coin summary to separate component and add coin summary skeleton
